+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## Versions
 
+- [`0.2.0`](#020)
+
+## 0.1.0
+
+*2021-02-05*
+
+### Major changes
+
+- [x] The auth method return itself ([#4](https://github.com/benjaminnoufel/node-oauth2-client/pull/4)).
+
+### Minor changes
+
+- [x] Having a getCredential method ([#4](https://github.com/benjaminnoufel/node-oauth2-client/pull/4)).
+- [x] Adding more examples ([#4](https://github.com/benjaminnoufel/node-oauth2-client/pull/4)).
+
+### Bug & security fixes
+
+None.
+
 - [`0.1.0`](#010)
 
 ## 0.1.0